wip: create apply subpackage
