+++ internal/cmd/marketplace.go
@@ -18,6 +18,7 @@ package cmd
 import (
 	"github.com/mia-platform/miactl/internal/clioptions"
 	"github.com/mia-platform/miactl/internal/cmd/marketplace"
+	marketplace_apply "github.com/mia-platform/miactl/internal/cmd/marketplace/apply"
 	"github.com/spf13/cobra"
 )
 
@@ -37,7 +38,7 @@ func MarketplaceCmd(options *clioptions.CLIOptions) *cobra.Command {
 	cmd.AddCommand(marketplace.ListCmd(options))
 	cmd.AddCommand(marketplace.GetCmd(options))
 	cmd.AddCommand(marketplace.DeleteCmd(options))
-	cmd.AddCommand(marketplace.ApplyCmd(options))
+	cmd.AddCommand(marketplace_apply.ApplyCmd(options))
 
 	return cmd
 }
