+++ internal/cmd/marketplace/apply/testdata/invalidJson1.json
