+++ internal/cmd/marketplace/apply/testdata/invalidYaml.yaml
