+++ internal/cmd/marketplace/apply/testdata/someFileNotYamlNotJson.txt
