+++ internal/cmd/marketplace/apply/testdata/validItem1.json
