+++ internal/cmd/marketplace/apply/testdata/validYaml.yaml
