+++ internal/cmd/marketplace/apply/testdata/validYaml.yml
