+++ internal/cmd/marketplace/apply/testdata/withoutErrors/validItem1.json
