+++ internal/cmd/marketplace/apply/testdata/withoutErrors/validYaml.yaml
